Remove Anchor obj.

+++ gameengine/src/Program.cpp
@@ -53,7 +53,7 @@ void Program::setUniform( const String& name, float value )
 
 void Program::setUniform( const String& name, unsigned value )
 {
-    getDevice()->useProgram( m_id );
+    getDevice()->useProgram( static_cast<int>( m_id ) );
 
     auto location = getUniformLocation( name );
     getDevice()->setUniformValue( location, value );
